Apply feeds
