--- src/ui/MessageBody/index.scss
@@ -1,9 +1,0 @@
-.sendbird-message-content__middle__message-item-body {
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.sendbird-voice-message-item-body.sendbird-message-content__middle__message-item-body {
-  min-width: 136px;
-  height: 50px;
-}
