--- src/ui/MessageProfile/index.scss
@@ -1,64 +1,0 @@
-.sendbird-message-content.incoming {
-  .sendbird-message-content__left {
-    position: relative;
-    display: inline-flex;
-    min-width: 40px;
-
-    .sendbird-message-content__left__avatar {
-      position: absolute;
-      left: 0px;
-      bottom: 2px;
-    }
-
-    .sendbird-message-content__left__avatar.use-thread-replies {
-      bottom: 35px;
-    }
-  }
-
-  .sendbird-message-content-menu {
-    position: relative;
-    flex-direction: row;
-    height: 32px;
-    display: none;
-    &.sendbird-mouse-hover {
-      display: inline-flex;
-    }
-  }
-}
-
-.sendbird-message-content.outgoing {
-  .sendbird-message-content__left {
-    position: relative;
-    box-sizing: border-box;
-    display: inline-flex;
-    justify-content: flex-end;
-    width: 50px;
-
-    &.use-reactions {
-      width: 70px
-    }
-
-    &.use-quote {
-      .sendbird-message-content-menu {
-        top: 18px;
-      }
-    }
-    .sendbird-message-content-menu {
-      position: absolute;
-      top: 2px;
-      right: 4px;
-      flex-direction: row;
-      display: none;
-      &.sendbird-mouse-hover {
-        display: inline-flex;
-      }
-    }
-  }
-  &:hover {
-    .sendbird-message-content__left {
-      .sendbird-message-content-menu {
-        display: inline-flex;
-      }
-    }
-  }
-}
